src/minerva_trigger_function/requirements.txt
# Function dependencies, for example:
# package>=version
google-auth==2.3.3
requests==2.26.0
